feat(geo_core): Vector3D完全実装をgeo_coreに移行、演算子問題を完全解決 (Issue #166 Phase 2完了)
- geo_core::Vector3D に完全実装を統合
  - 基本機能（new, zero, unit_x/y/z, dot, cross等）
  - Foundation トレイト実装完全対応
  - 拡張機能（lerp, slerp, project_onto, reflect等）
- Point3D と Vector3D の演算子オーバーロードを geo_core で実装
  - Point - Point = Vector
  - Point + Vector = Point
  - Point - Vector = Point
  - Vector ± Vector, Vector × Scalar 等
- geo_primitives は pub use geo_core::{Point3D, Vector3D} のみに変更
- vector_3d_extensions.rs, vector_3d_foundation.rs 削除

**完了事項**:
- オーファンルール問題を完全解決
- 演算子オーバーロードにより既存コード（p1 - p2, center + v）がそのまま動作
- 全テスト成功: 704 tests passed

**アーキテクチャ**:
`
geo_foundation (トレイト定義)
    ↓
geo_core (Point3D + Vector3D 完全実装 + 演算子)
    ↓ pub use
geo_primitives (再エクスポートのみ)
`

geo_core が geo_primitives の下位層として、Point3D と Vector3D の完全実装を提供。
この構成により、型安全性とアーキテクチャの明確性を両立。
