+++ model/geo_primitives/src/point_3d.rs
@@ -1,4 +1,8 @@
 //! Point3D - geo_core からの再エクスポート
 
+// ============================================================================
+// Re-export from geo_core
+// ============================================================================
+
 /// geo_core::Point3D を再エクスポート
 pub use geo_core::Point3D;
